FIXED: bug in PUT DATA.
As a consequence of this fix, the applet now runs fine in java card kit CREF emulator (except for AA, still have to fix that).

+++ passportapplet/src/sos/passportapplet/PassportUtil.java
@@ -149,7 +149,7 @@ public class PassportUtil implements ISO7816 {
     }
     
     public static short lengthWithPadding(short inputLength) {
-    	return (short)(((inputLength + 8) / 8) * 8); 
+    	return (short)((((short)(inputLength + 8)) / 8) * 8); 
     }
     
 //    public static void pad(APDU aapdu, short pad_len) {
